Chnaged the type of LastUpdated property form int? to long? in EpisodeRecord, Series and Update models

+++ src/TvDbSharper/Clients/Episodes/Json/EpisodeRecord.cs
@@ -41,7 +41,7 @@
 
         public string ImdbId { get; set; }
 
-        public int? LastUpdated { get; set; }
+        public long? LastUpdated { get; set; }
 
         public string LastUpdatedBy { get; set; }
